+++ DalaMock.Host/Mediator/DisposableMediatorSubscriberBase.cs
@@ -1,12 +1,13 @@
 namespace DalaMock.Host.Mediator;
 
 using System;
-using Dalamud.Plugin.Services;
+
+using Microsoft.Extensions.Logging;
 
 public abstract class DisposableMediatorSubscriberBase : MediatorSubscriberBase, IDisposable
 {
     protected DisposableMediatorSubscriberBase(
-        IPluginLog logger,
+        ILogger<DisposableMediatorSubscriberBase> logger,
         MediatorService
             mediatorService)
         : base(logger, mediatorService)
@@ -21,7 +22,7 @@ public abstract class DisposableMediatorSubscriberBase : MediatorSubscriberBase,
 
     protected virtual void Dispose(bool disposing)
     {
-        this.Logger.Verbose("Disposing {type} ({this})", this.GetType().Name, this);
+        this.Logger.LogTrace("Disposing {type} ({this})", this.GetType().Name, this);
         this.UnsubscribeAll();
     }
 }