+++ DalaMock.Host/Mediator/MediatorSubscriberBase.cs
@@ -1,23 +1,24 @@
 namespace DalaMock.Host.Mediator;
 
 using Dalamud.Plugin.Services;
+using Microsoft.Extensions.Logging;
 
 public abstract class MediatorSubscriberBase : IMediatorSubscriber
 {
-    protected MediatorSubscriberBase(IPluginLog logger, MediatorService mediatorService)
+    protected MediatorSubscriberBase(ILogger<MediatorSubscriberBase> logger, MediatorService mediatorService)
     {
         this.Logger = logger;
-        this.Logger.Verbose("Creating {type} ({this})", this.GetType().Name, this);
+        this.Logger.LogTrace("Creating {type} ({this})", this.GetType().Name, this);
         this.MediatorService = mediatorService;
     }
 
-    protected IPluginLog Logger { get; }
+    protected ILogger Logger { get; }
 
     public MediatorService MediatorService { get; }
 
     protected void UnsubscribeAll()
     {
-        this.Logger.Verbose("Unsubscribing from all for {type} ({this})", this.GetType().Name, this);
+        this.Logger.LogTrace("Unsubscribing from all for {type} ({this})", this.GetType().Name, this);
         this.MediatorService.UnsubscribeAll(this);
     }
 }